modules/location_generator/requirements.txt
grpcio-tools == 1.51.1
grpcio == 1.51.1
kafka-python == 2.0.2